requirements.txt
streamlit
supabase
Pillow
opencv-python-headless